// test/types/options.schema.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface OptionsSchema {
  schema: {
    [k: string]: any;
  };
  data?:
    | [
        {
          [k: string]: any;
        },
        ...{
          [k: string]: any;
        }[]
      ]
    | {
        [k: string]: any;
      };
  env?: boolean;
  dotenv?:
    | boolean
    | {
        [k: string]: any;
      };
  [k: string]: any;
}